requirements.txt
fastapi>=0.115.2
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
Levenshtein>=0.21.0
requests>=2.31.0
python-multipart>=0.0.6
starlette>=0.40.0
groq>=0.11.0
piper-tts>=1.2.0
onnxruntime>=1.15.0
numpy<2.0.0